long -> int to fix 64bit issues.  This breaks 16bit machines but since we don't have any 16bit ports...) The alternative is to use int32_t's but then this code will break when/if int > 32bits which is not so unlikely.  Fixes NetBSD PR #3450 from Ross Harvey (but in a different way)

+++ lib/libc/regex/regex2.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: regex2.h,v 1.5 1995/02/27 13:29:40 cgd Exp $	*/
+/*	$OpenBSD: regex2.h,v 1.2 1996/08/19 08:31:14 tholo Exp $	*/
 
 /*-
  * Copyright (c) 1992, 1993, 1994 Henry Spencer.
@@ -75,8 +75,8 @@
  * In state representations, an operator's bit is on to signify a state
  * immediately *preceding* "execution" of that operator.
  */
-typedef unsigned long sop;	/* strip operator */
-typedef long sopno;
+typedef unsigned sop;	/* strip operator */
+typedef int sopno;
 #define	OPRMASK	0xf8000000
 #define	OPDMASK	0x07ffffff
 #define	OPSHIFT	((unsigned)27)
